+++ orderplanning/src/main/java/com/overc1ock/service/ContractService.java
@@ -11,8 +11,8 @@ public interface ContractService {
 	List<ContractVO> getItemCodeForContract(); // 품목 코드 가져오기
 	List<ContractVO> getSubcontractorName(); // 협력 업체명 가져오기
 	void registerContract(ContractVO vo); // 등록
-	ContractVO inquiryContract(Long contract_code); // 조회
-	boolean deleteContract(Long contract_code); // 삭제
+	ContractVO inquiryContract(Integer contract_code); // 조회
+	boolean deleteContract(Integer contract_code); // 삭제
 	boolean modifyContract(ContractVO vo); // 수정
 
 }
